% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre_parity.R
\name{caller_file_runtime}
\alias{caller_file_runtime}
\title{Get the Runtime Caller File}
\usage{
caller_file_runtime(hint_path = NULL)
}
\arguments{
\item{hint_path}{Optional hint path.}
}
\value{
Caller file path.
}
\description{
Get the Runtime Caller File
}
