% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre_parity.R
\name{canonical_path}
\alias{canonical_path}
\title{Canonicalize a Filesystem Path}
\usage{
canonical_path(path)
}
\arguments{
\item{path}{Filesystem path.}
}
\value{
Canonicalized path.
}
\description{
Canonicalize a Filesystem Path
}
