% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre_parity.R
\name{file_uri_to_path}
\alias{file_uri_to_path}
\title{Convert a File URI to a Path}
\usage{
file_uri_to_path(uri)
}
\arguments{
\item{uri}{File URI string.}
}
\value{
A local filesystem path.
}
\description{
Convert a File URI to a Path
}
