% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre_parity.R
\name{get_datafilename}
\alias{get_datafilename}
\title{Build a TRE Data Filename}
\usage{
get_datafilename(asset_name, major = 1L, minor = 0L, patch = 0L)
}
\arguments{
\item{asset_name}{Asset identifier.}

\item{major}{Major version number.}

\item{minor}{Minor version number.}

\item{patch}{Patch version number.}
}
\value{
Filename string.
}
\description{
Build a TRE Data Filename
}
