% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre_parity.R
\name{get_datalake_file_path}
\alias{get_datalake_file_path}
\title{Build a Datalake File Path}
\usage{
get_datalake_file_path(
  lake_data,
  study_name,
  asset_name,
  source_file_path,
  major = 1L,
  minor = 0L,
  patch = 0L
)
}
\arguments{
\item{lake_data}{Root lake data path.}

\item{study_name}{Study identifier.}

\item{asset_name}{Asset identifier.}

\item{source_file_path}{Source filename or path.}

\item{major}{Major version number.}

\item{minor}{Minor version number.}

\item{patch}{Patch version number.}
}
\value{
Datalake file path.
}
\description{
Build a Datalake File Path
}
