% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre_parity.R
\name{get_datasetname}
\alias{get_datasetname}
\title{Build a TRE Dataset Name}
\usage{
get_datasetname(
  study_name,
  asset_name,
  major = 1L,
  minor = 0L,
  patch = 0L,
  include_schema = FALSE
)
}
\arguments{
\item{study_name}{Study identifier.}

\item{asset_name}{Asset identifier.}

\item{major}{Major version number.}

\item{minor}{Minor version number.}

\item{patch}{Patch version number.}

\item{include_schema}{Whether to include schema qualification.}
}
\value{
Dataset name string.
}
\description{
Build a TRE Dataset Name
}
