% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHRITREC.R
\name{get_redcap_choices_for_field}
\alias{get_redcap_choices_for_field}
\title{Get REDCap field choices}
\usage{
get_redcap_choices_for_field(field_type, choices = NULL)
}
\arguments{
\item{field_type}{REDCap field type.}

\item{choices}{Optional REDCap choices string.}
}
\value{
A JSON string describing the available field choices.
}
\description{
Convenience wrapper around \code{\link[=get_redcap_choices_for_field_json]{get_redcap_choices_for_field_json()}} that
returns the same JSON payload.
}
