% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHRITREC.R
\name{infer_label_from_field_name}
\alias{infer_label_from_field_name}
\title{Infer a human-readable label from a field name}
\usage{
infer_label_from_field_name(field_name)
}
\arguments{
\item{field_name}{Field name to transform.}
}
\value{
A human-readable label string.
}
\description{
Converts a machine-oriented field name into a more readable label.
}
