% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre_parity.R
\name{is_ncname}
\alias{is_ncname}
\title{Check XML NCName Validity}
\usage{
is_ncname(value, strict = TRUE)
}
\arguments{
\item{value}{Candidate value.}

\item{strict}{Whether to apply strict validation rules.}
}
\value{
\code{TRUE} when \code{value} is a valid NCName.
}
\description{
Check XML NCName Validity
}
