% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre_parity.R
\name{julia_type_to_sql_string}
\alias{julia_type_to_sql_string}
\title{Map a Julia Type Name to SQL}
\usage{
julia_type_to_sql_string(julia_type)
}
\arguments{
\item{julia_type}{Julia type name.}
}
\value{
SQL type string.
}
\description{
Map a Julia Type Name to SQL
}
