% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre_parity.R
\name{makeparams}
\alias{makeparams}
\title{Create Positional SQL Parameters}
\usage{
makeparams(n)
}
\arguments{
\item{n}{Number of parameters.}
}
\value{
JSON string representing placeholder parameters.
}
\description{
Parity alias for \code{makeparams_json()}.
}
