% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre_parity.R
\name{makeparams_json}
\alias{makeparams_json}
\title{Create Positional SQL Parameters as JSON}
\usage{
makeparams_json(n)
}
\arguments{
\item{n}{Number of parameters.}
}
\value{
JSON string representing placeholder parameters.
}
\description{
Create Positional SQL Parameters as JSON
}
