% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHRITREC.R
\name{map_sql_type_to_tre}
\alias{map_sql_type_to_tre}
\title{Map a SQL type to a TRE type}
\usage{
map_sql_type_to_tre(sql_type)
}
\arguments{
\item{sql_type}{SQL type text to map.}
}
\value{
A TRE type string.
}
\description{
Converts a SQL column type into the TRE type name used by AHRI TRE.
}
