% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre_parity.R
\name{map_sql_type_to_tre_for_flavour}
\alias{map_sql_type_to_tre_for_flavour}
\title{Map SQL Type to TRE Type for a Flavour}
\usage{
map_sql_type_to_tre_for_flavour(sql_type, flavour_id)
}
\arguments{
\item{sql_type}{SQL type name.}

\item{flavour_id}{A value from \code{DatabaseFlavour}.}
}
\value{
Integer TRE type identifier.
}
\description{
Map SQL Type to TRE Type for a Flavour
}
