% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre_parity.R
\name{normalise_orcid_rolename}
\alias{normalise_orcid_rolename}
\title{Normalize an ORCID Role Name}
\usage{
normalise_orcid_rolename(orcid)
}
\arguments{
\item{orcid}{ORCID or ORCID-based role string.}
}
\value{
Normalized role name.
}
\description{
Normalize an ORCID Role Name
}
