% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre_parity.R
\name{normalize_git_remote}
\alias{normalize_git_remote}
\alias{_normalize_remote}
\title{Normalize a Git Remote URL}
\usage{
normalize_git_remote(url)

`_normalize_remote`(url)
}
\arguments{
\item{url}{Remote URL.}
}
\value{
Normalized remote string.
}
\description{
Normalize a Git Remote URL
}
