% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHRITREC.R
\name{parse_in_list_values}
\alias{parse_in_list_values}
\title{Parse \code{IN (...)} values}
\usage{
parse_in_list_values(values)
}
\arguments{
\item{values}{Text containing values to parse.}
}
\value{
A JSON string describing the parsed values.
}
\description{
Convenience wrapper around \code{\link[=parse_in_list_values_json]{parse_in_list_values_json()}} that returns the
same JSON payload.
}
