% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHRITREC.R
\name{parse_in_list_values_json}
\alias{parse_in_list_values_json}
\title{Parse \code{IN (...)} values as JSON}
\usage{
parse_in_list_values_json(values)
}
\arguments{
\item{values}{Text containing values to parse.}
}
\value{
A JSON string describing the parsed values.
}
\description{
Parses a string containing SQL-style \code{IN (...)} values and returns the parsed
result as JSON text.
}
