% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHRITREC.R
\name{parse_redcap_choices}
\alias{parse_redcap_choices}
\title{Parse REDCap choices}
\usage{
parse_redcap_choices(choices)
}
\arguments{
\item{choices}{REDCap choices string.}
}
\value{
A JSON string describing the parsed choices.
}
\description{
Convenience wrapper around \code{\link[=parse_redcap_choices_json]{parse_redcap_choices_json()}} that returns the
same JSON payload.
}
