% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHRITREC.R
\name{parse_redcap_choices_json}
\alias{parse_redcap_choices_json}
\title{Parse REDCap choices as JSON}
\usage{
parse_redcap_choices_json(choices)
}
\arguments{
\item{choices}{REDCap choices string.}
}
\value{
A JSON string describing the parsed choices.
}
\description{
Parses REDCap \verb{choices, calculations, OR slider labels} text into JSON.
}
