% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre_parity.R
\name{path_to_file_uri}
\alias{path_to_file_uri}
\title{Convert a Path to a File URI}
\usage{
path_to_file_uri(path)
}
\arguments{
\item{path}{Local filesystem path.}
}
\value{
A file URI string.
}
\description{
Convert a Path to a File URI
}
