% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre_parity.R
\name{quote_ident}
\alias{quote_ident}
\title{Quote a SQL Identifier}
\usage{
quote_ident(name)
}
\arguments{
\item{name}{Identifier to quote.}
}
\value{
Quoted identifier.
}
\description{
Quote a SQL Identifier
}
