% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre_parity.R
\name{quote_identifier}
\alias{quote_identifier}
\title{Quote an Identifier}
\usage{
quote_identifier(name)
}
\arguments{
\item{name}{Identifier to quote.}
}
\value{
Quoted identifier.
}
\description{
Parity alias for \code{quote_ident()}.
}
