% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre_parity.R
\name{quote_qualified_identifier}
\alias{quote_qualified_identifier}
\title{Quote a Qualified SQL Identifier}
\usage{
quote_qualified_identifier(name)
}
\arguments{
\item{name}{Qualified identifier to quote.}
}
\value{
Quoted identifier.
}
\description{
Quote a Qualified SQL Identifier
}
