% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre_parity.R
\name{quote_sql_str}
\alias{quote_sql_str}
\title{Quote a SQL String Literal}
\usage{
quote_sql_str(value)
}
\arguments{
\item{value}{String literal to quote.}
}
\value{
Quoted SQL string literal.
}
\description{
Quote a SQL String Literal
}
