% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tre_parity.R
\name{tre_type_to_duckdb_sql}
\alias{tre_type_to_duckdb_sql}
\title{Map a TRE Type to DuckDB SQL}
\usage{
tre_type_to_duckdb_sql(value_type_id)
}
\arguments{
\item{value_type_id}{TRE type identifier.}
}
\value{
DuckDB SQL type string.
}
\description{
Map a TRE Type to DuckDB SQL
}
