% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHRITREC.R
\name{version}
\alias{version}
\title{Return the core library version}
\usage{
version()
}
\value{
A single character string containing the library version.
}
\description{
Calls into the AHRI TRE C core and returns its version string.
}
