\name{tre_requested_api_stubs}
\alias{ColumnInfo}
\alias{add_datastore_orcid}
\alias{add_domain}
\alias{add_entity}
\alias{add_entity_relation}
\alias{add_study}
\alias{add_study_domain}
\alias{add_transformation}
\alias{add_transformation_input}
\alias{add_transformation_output}
\alias{add_variable}
\alias{attach_datafile}
\alias{attach_datafile_version}
\alias{closedatastore}
\alias{connect_mssql}
\alias{convert_missing_to_string}
\alias{create_asset}
\alias{create_dataset_meta}
\alias{create_duckdb_table_sql}
\alias{create_lake_database}
\alias{create_store_database}
\alias{create_transformation}
\alias{createassets}
\alias{createdatastore}
\alias{createentities}
\alias{createmapping}
\alias{createstudies}
\alias{createtransformations}
\alias{createvariables}
\alias{dataset_to_arrow}
\alias{dataset_to_csv}
\alias{dataset_to_dataframe}
\alias{dataset_variables}
\alias{ensure_mssql_driver_registered}
\alias{ensure_vocabulary}
\alias{find_mssql_driver_in_directory}
\alias{find_system_odbc_driver}
\alias{get_asset}
\alias{get_assetversions}
\alias{get_code_table_vocabulary}
\alias{get_column_comment}
\alias{get_column_type_info}
\alias{get_datafile}
\alias{get_datafile_meta}
\alias{get_datafile_metadata}
\alias{get_dataset}
\alias{get_dataset_variables}
\alias{get_dataset_versions}
\alias{get_domain}
\alias{get_domain_variables}
\alias{get_domainentities}
\alias{get_domainrelations}
\alias{get_domains}
\alias{get_eav_variable_names}
\alias{get_entity}
\alias{get_entityrelation}
\alias{get_enum_values}
\alias{get_foreign_key_reference}
\alias{get_latest_version}
\alias{get_namedkey}
\alias{get_original_column_type}
\alias{get_query_columns}
\alias{get_studies}
\alias{get_study}
\alias{get_study_assets}
\alias{get_study_datafiles}
\alias{get_study_datasets}
\alias{get_study_domains}
\alias{get_study_variables}
\alias{get_study_variables_df}
\alias{get_studyid}
\alias{get_table}
\alias{get_table_columns}
\alias{get_variable}
\alias{get_variable_id}
\alias{get_vocabularies}
\alias{get_vocabulary}
\alias{ingest_file}
\alias{ingest_file_version}
\alias{ingest_redcap_project}
\alias{initstudytypes}
\alias{initvalue_types}
\alias{insertdata}
\alias{insertwithidentity}
\alias{is_code_table}
\alias{is_enum_type}
\alias{lines}
\alias{list_domainentities}
\alias{list_domainrelations}
\alias{list_study_assets_df}
\alias{list_study_transformations}
\alias{load_query}
\alias{make_asset}
\alias{opendatastore}
\alias{prepareinsertstatement}
\alias{prepareselectstatement}
\alias{read_dataset}
\alias{redcap_export_eav}
\alias{redcap_fields}
\alias{redcap_metadata}
\alias{redcap_post}
\alias{redcap_post_tofile}
\alias{redcap_project_info}
\alias{redcap_project_info_df}
\alias{register_datafile}
\alias{register_dataset}
\alias{register_redcap_datadictionary}
\alias{save_asset_version}
\alias{save_dataset_variables}
\alias{save_version}
\alias{savedataframe}
\alias{selectdataframe}
\alias{set_version}
\alias{sql_meta}
\alias{sql_to_dataset}
\alias{table_exists}
\alias{table_has_primary_key}
\alias{transaction_begin}
\alias{transaction_commit}
\alias{transaction_rollback}
\alias{transform_eav_to_dataset}
\alias{transform_eav_to_table}
\alias{update_domain}
\alias{update_variable}
\alias{updatevalues}
\alias{upsert_entity}
\alias{upsert_entityrelation}
\alias{upsert_study}
\alias{upsert_variable}
\alias{vocabulary_items}
\alias{wrap_query_for_metadata}
\title{Compatibility Placeholder APIs}
\description{
Compatibility placeholder functions exported for API parity with AHRI TRE workflows.
These entry points currently exist so callers can resolve the names from the
package namespace, but they are not yet implemented in the R package.
}
\usage{
ColumnInfo(...)
add_datastore_orcid(...)
add_domain(...)
add_entity(...)
add_entity_relation(...)
add_study(...)
add_study_domain(...)
add_transformation(...)
add_transformation_input(...)
add_transformation_output(...)
add_variable(...)
attach_datafile(...)
attach_datafile_version(...)
closedatastore(...)
connect_mssql(...)
convert_missing_to_string(...)
create_asset(...)
create_dataset_meta(...)
create_duckdb_table_sql(...)
create_lake_database(...)
create_store_database(...)
create_transformation(...)
createassets(...)
createdatastore(...)
createentities(...)
createmapping(...)
createstudies(...)
createtransformations(...)
createvariables(...)
dataset_to_arrow(...)
dataset_to_csv(...)
dataset_to_dataframe(...)
dataset_variables(...)
ensure_mssql_driver_registered(...)
ensure_vocabulary(...)
find_mssql_driver_in_directory(...)
find_system_odbc_driver(...)
get_asset(...)
get_assetversions(...)
get_code_table_vocabulary(...)
get_column_comment(...)
get_column_type_info(...)
get_datafile(...)
get_datafile_meta(...)
get_datafile_metadata(...)
get_dataset(...)
get_dataset_variables(...)
get_dataset_versions(...)
get_domain(...)
get_domain_variables(...)
get_domainentities(...)
get_domainrelations(...)
get_domains(...)
get_eav_variable_names(...)
get_entity(...)
get_entityrelation(...)
get_enum_values(...)
get_foreign_key_reference(...)
get_latest_version(...)
get_namedkey(...)
get_original_column_type(...)
get_query_columns(...)
get_studies(...)
get_study(...)
get_study_assets(...)
get_study_datafiles(...)
get_study_datasets(...)
get_study_domains(...)
get_study_variables(...)
get_study_variables_df(...)
get_studyid(...)
get_table(...)
get_table_columns(...)
get_variable(...)
get_variable_id(...)
get_vocabularies(...)
get_vocabulary(...)
ingest_file(...)
ingest_file_version(...)
ingest_redcap_project(...)
initstudytypes(...)
initvalue_types(...)
insertdata(...)
insertwithidentity(...)
is_code_table(...)
is_enum_type(...)
lines(...)
list_domainentities(...)
list_domainrelations(...)
list_study_assets_df(...)
list_study_transformations(...)
load_query(...)
make_asset(...)
opendatastore(...)
prepareinsertstatement(...)
prepareselectstatement(...)
read_dataset(...)
redcap_export_eav(...)
redcap_fields(...)
redcap_metadata(...)
redcap_post(...)
redcap_post_tofile(...)
redcap_project_info(...)
redcap_project_info_df(...)
register_datafile(...)
register_dataset(...)
register_redcap_datadictionary(...)
save_asset_version(...)
save_dataset_variables(...)
save_version(...)
savedataframe(...)
selectdataframe(...)
set_version(...)
sql_meta(...)
sql_to_dataset(...)
table_exists(...)
table_has_primary_key(...)
transaction_begin(...)
transaction_commit(...)
transaction_rollback(...)
transform_eav_to_dataset(...)
transform_eav_to_table(...)
update_domain(...)
update_variable(...)
updatevalues(...)
upsert_entity(...)
upsert_entityrelation(...)
upsert_study(...)
upsert_variable(...)
vocabulary_items(...)
wrap_query_for_metadata(...)
}
\arguments{
\item{...}{Unused arguments. These compatibility placeholders currently stop with a not-yet-implemented error.}
}
\value{
These functions do not currently return a value. They abort with an informative
error until a concrete implementation is added.
}% Generated by helper script: compatibility docs for placeholder API
\name{tre_requested_api_stubs}
\alias{ColumnInfo}
\alias{add_datastore_orcid}
\alias{add_domain}
\alias{add_entity}
\alias{add_entity_relation}
\alias{add_study}
\alias{add_study_domain}
\alias{add_transformation}
\alias{add_transformation_input}
\alias{add_transformation_output}
\alias{add_variable}
\alias{attach_datafile}
\alias{attach_datafile_version}
\alias{closedatastore}
\alias{connect_mssql}
\alias{convert_missing_to_string}
\alias{create_asset}
\alias{create_dataset_meta}
\alias{create_duckdb_table_sql}
\alias{create_lake_database}
\alias{create_store_database}
\alias{create_transformation}
\alias{createassets}
\alias{createdatastore}
\alias{createentities}
\alias{createmapping}
\alias{createstudies}
\alias{createtransformations}
\alias{createvariables}
\alias{dataset_to_arrow}
\alias{dataset_to_csv}
\alias{dataset_to_dataframe}
\alias{dataset_variables}
\alias{ensure_mssql_driver_registered}
\alias{ensure_vocabulary}
\alias{find_mssql_driver_in_directory}
\alias{find_system_odbc_driver}
\alias{get_asset}
\alias{get_assetversions}
\alias{get_code_table_vocabulary}
\alias{get_column_comment}
\alias{get_column_type_info}
\alias{get_datafile}
\alias{get_datafile_meta}
\alias{get_datafile_metadata}
\alias{get_dataset}
\alias{get_dataset_variables}
\alias{get_dataset_versions}
\alias{get_domain}
\alias{get_domain_variables}
\alias{get_domainentities}
\alias{get_domainrelations}
\alias{get_domains}
\alias{get_eav_variable_names}
\alias{get_entity}
\alias{get_entityrelation}
\alias{get_enum_values}
\alias{get_foreign_key_reference}
\alias{get_latest_version}
\alias{get_namedkey}
\alias{get_original_column_type}
\alias{get_query_columns}
\alias{get_studies}
\alias{get_study}
\alias{get_study_assets}
\alias{get_study_datafiles}
\alias{get_study_datasets}
\alias{get_study_domains}
\alias{get_study_variables}
\alias{get_study_variables_df}
\alias{get_studyid}
\alias{get_table}
\alias{get_table_columns}
\alias{get_variable}
\alias{get_variable_id}
\alias{get_vocabularies}
\alias{get_vocabulary}
\alias{ingest_file}
\alias{ingest_file_version}
\alias{ingest_redcap_project}
\alias{initstudytypes}
\alias{initvalue_types}
\alias{insertdata}
\alias{insertwithidentity}
\alias{is_code_table}
\alias{is_enum_type}
\alias{lines}
\alias{list_domainentities}
\alias{list_domainrelations}
\alias{list_study_assets_df}
\alias{list_study_transformations}
\alias{load_query}
\alias{make_asset}
\alias{opendatastore}
\alias{prepareinsertstatement}
\alias{prepareselectstatement}
\alias{read_dataset}
\alias{redcap_export_eav}
\alias{redcap_fields}
\alias{redcap_metadata}
\alias{redcap_post}
\alias{redcap_post_tofile}
\alias{redcap_project_info}
\alias{redcap_project_info_df}
\alias{register_datafile}
\alias{register_dataset}
\alias{register_redcap_datadictionary}
\alias{save_asset_version}
\alias{save_dataset_variables}
\alias{save_version}
\alias{savedataframe}
\alias{selectdataframe}
\alias{set_version}
\alias{sql_meta}
\alias{sql_to_dataset}
\alias{table_exists}
\alias{table_has_primary_key}
\alias{transaction_begin}
\alias{transaction_commit}
\alias{transaction_rollback}
\alias{transform_eav_to_dataset}
\alias{transform_eav_to_table}
\alias{update_domain}
\alias{update_variable}
\alias{updatevalues}
\alias{upsert_entity}
\alias{upsert_entityrelation}
\alias{upsert_study}
\alias{upsert_variable}
\alias{vocabulary_items}
\alias{wrap_query_for_metadata}
\title{Compatibility API Placeholders}
\arguments{
\item{...}{Placeholder arguments reserved for future implementations.}
}
\usage{
ColumnInfo(...)
add_datastore_orcid(...)
add_domain(...)
add_entity(...)
add_entity_relation(...)
add_study(...)
add_study_domain(...)
add_transformation(...)
add_transformation_input(...)
add_transformation_output(...)
add_variable(...)
attach_datafile(...)
attach_datafile_version(...)
closedatastore(...)
connect_mssql(...)
convert_missing_to_string(...)
create_asset(...)
create_dataset_meta(...)
create_duckdb_table_sql(...)
create_lake_database(...)
create_store_database(...)
create_transformation(...)
createassets(...)
createdatastore(...)
createentities(...)
createmapping(...)
createstudies(...)
createtransformations(...)
createvariables(...)
dataset_to_arrow(...)
dataset_to_csv(...)
dataset_to_dataframe(...)
dataset_variables(...)
ensure_mssql_driver_registered(...)
ensure_vocabulary(...)
find_mssql_driver_in_directory(...)
find_system_odbc_driver(...)
get_asset(...)
get_assetversions(...)
get_code_table_vocabulary(...)
get_column_comment(...)
get_column_type_info(...)
get_datafile(...)
get_datafile_meta(...)
get_datafile_metadata(...)
get_dataset(...)
get_dataset_variables(...)
get_dataset_versions(...)
get_domain(...)
get_domain_variables(...)
get_domainentities(...)
get_domainrelations(...)
get_domains(...)
get_eav_variable_names(...)
get_entity(...)
get_entityrelation(...)
get_enum_values(...)
get_foreign_key_reference(...)
get_latest_version(...)
get_namedkey(...)
get_original_column_type(...)
get_query_columns(...)
get_studies(...)
get_study(...)
get_study_assets(...)
get_study_datafiles(...)
get_study_datasets(...)
get_study_domains(...)
get_study_variables(...)
get_study_variables_df(...)
get_studyid(...)
get_table(...)
get_table_columns(...)
get_variable(...)
get_variable_id(...)
get_vocabularies(...)
get_vocabulary(...)
ingest_file(...)
ingest_file_version(...)
ingest_redcap_project(...)
initstudytypes(...)
initvalue_types(...)
insertdata(...)
insertwithidentity(...)
is_code_table(...)
is_enum_type(...)
lines(...)
list_domainentities(...)
list_domainrelations(...)
list_study_assets_df(...)
list_study_transformations(...)
load_query(...)
make_asset(...)
opendatastore(...)
prepareinsertstatement(...)
prepareselectstatement(...)
read_dataset(...)
redcap_export_eav(...)
redcap_fields(...)
redcap_metadata(...)
redcap_post(...)
redcap_post_tofile(...)
redcap_project_info(...)
redcap_project_info_df(...)
register_datafile(...)
register_dataset(...)
register_redcap_datadictionary(...)
save_asset_version(...)
save_dataset_variables(...)
save_version(...)
savedataframe(...)
selectdataframe(...)
set_version(...)
sql_meta(...)
sql_to_dataset(...)
table_exists(...)
table_has_primary_key(...)
transaction_begin(...)
transaction_commit(...)
transaction_rollback(...)
transform_eav_to_dataset(...)
transform_eav_to_table(...)
update_domain(...)
update_variable(...)
updatevalues(...)
upsert_entity(...)
upsert_entityrelation(...)
upsert_study(...)
upsert_variable(...)
vocabulary_items(...)
wrap_query_for_metadata(...)
}
\description{
Compatibility placeholders for requested API surface. These functions currently exist for interface parity and may stop with a not-yet-implemented message when called.
}
\details{
These entries provide documentation coverage for exported compatibility symbols while implementations are phased in.
}
